requirements.txt
blind-watermark>=0.4.0
opencv-python>=4.5.0
numpy<2.0.0
click>=8.0.0
Pillow>=9.0.0
pyfiglet
rich
python-pptx
questionary
cryptography